PeerReviewErrors.py
Reviewing other developer’s work is essential. Reviewing the work of others makes both the reviewer and the reviewee better developers by allowing you to see how others might solve the same problem.  It also ensures that standards are adhered to and offer the potential to catch logic errors that would otherwise end up in production.

For this assignment, you will peer review this sample code (click the link to download a .zip file that contains a .py file).